stream: add inline Game_StateSetNext

+++ src/screens/credits/credits.c
@@ -171,7 +171,7 @@ s32 func_801E3304(void)
             g_SysWork.field_14 = 0;
 
             g_GameWork.gameStateStep_598[0] = prevState;
-            g_GameWork.gameState_594        = (s32)g_GameWork.gameStatePrev_590;
+            g_GameWork.gameState_594        = g_GameWork.gameStatePrev_590;
             g_GameWork.gameStatePrev_590    = prevState;
             g_GameWork.gameStateStep_598[0] = 0;
         }
